[WIP] Remove unnecessary timed wait for completable future
Computation of completion proposals and getting auto activation
characters use the same CompletableFuture.

When processing the completion proposals gets delayed, and the auto
activation characters are requested via keyboard shortcut, there is a
timeout because of the wait for the completionLanguageServersFuture.

This wait has been removed because there is a cache that exists, and the
latest version will be returned, regardless of the delays.